AWS Explorer now refreshes when Credentials change - decoupled refresh from DefaultAWSContextCommands login/logout - awsconext no longer raises events for region changes
